RF/BF: move DANDISwaggerAutoSchema into dedicated swagger.py
